// src/components/PortfolioOverview/types.ts
export interface IWSAnswer {
  A: string;
  B: string;
  C: number;
  E: number;
  F: number;
  L: number;
  O: number;
  P: string;
  Q: string;
  a: string;
  b: string;
  c: string;
  e: string;
  h: string;
  l: string;
  n: number;
  o: string;
  p: string;
  q: string;
  s: string;
  v: string;
  w: string;
  x: string;
}

export interface IAsset {
  id: string;
  name: string;
  amount: number;
  price: number;
  totalValue: number;
  change24h: number;
  portfolioShare: number;
}